using type parameter for assertion functions
best practice of writing assertion functions is using a generic. This way I am able to restrict the type parameter of IAssertionFactory to invariant in the future without the need to change these functions. Making it invariant might be an interesting step in case partial type parameter inference is supported by Kotlin.