split project up into api, impl and verbs
This has two purposes: first, each project should define its own assertion verbs.
The provided verbs are only for the impatient and thus should not be in the same project.
Secondly, we want to be able to exchange the implementation and for this to work we need an API and a volatile API which will follow in a next commit