prefix for `describe` in specs
Assuming one might want to write further specs and not just use the
predefined specs provided by Atrium: if a `describe` has the same
description as a `describe` in the specs provided by Atrium, then
only one of the two are evaluated. In order to prevent name clashes
one would need to take care that one's `describe`s differ from the
Atrium's `describe`s.
To circumvent this situation one can define a prefix for `describe`
(defaults to "[Atrium] ") so that a) one does not have to think about
name clashes and b) it is obvious in the output what was defined by
Atrium and what are additional specifications of the implementation

- add a prefix to all atrium-spec's
- turn atrium-spec's into abstract classes (so that the spek-plugin
  does not try to execute it when we choose run all)