child assertion should append bulletPoint
Currently an assertion group formatter decides which bullet point shall
be used and appends the bullet point for the first of its assertions.
This leads to lisp like nesting of lists which makes it harder to
understand the program (see the need for indentIndex in
IIndentAssertionGroupType). Instead the assertion group formatter should
only suggest what a child assertion should use as bullet point but it is
up to the child assertion to decide if it is using it or not.

- move appendLn, indent and append prefix to the part where the name
  and subject of an assertion group, the assertion itself is formatted
  - do not appendLn for RootAssertionGroupType (fallback formatter)
  - adopt Specs for IAssertionFormatter; they all include an
    appendLn at the beginning (except they are in the root group)
- get rid of indentIndex in IIndentAssertionGroupType
- rename TextSameLineAssertionFormatterSpec (was forgotten when the
  class was renamed)