bullet points should be extendable
I think it is cumbersome to add a new parameter to
ReporterBuilder::withSameLineTextAssertionFormatter every time we get
a new IAssertionGroupType. Instead it should be open for extension (so
that also others could define more bullet points later on). Another
positive aspect we get out of this: the user will be able to define the
indentation for the different types (if desired).

- introduce marker interface IBulletPointIdentifier
  - make IAssertionGroupType a subtype of it
  - introduce subclass PrefixFeatureAssertionGroupHeader for the arrow
    of feature assertion groups
- use a Map instead of several Strings where IBulletPointIdentifier is
  used as key
  - modify IAtriumFactory accordingly
  - modify the constructors of the IAssertionFormatters accordingly
  - modify the specs due to constructor change
    => introduce AssertionFormatterSpecBase to reduce duplication
  - rewrite KDoc