introduce ISummaryAssertionGroupType
Sometimes we want to report a result as a whole (e.g. when using
`assert(setOf(1,2)).contains.inAnyOrder.only.values(1,2,3)`) and not
just the failing assertions (in the example the values which cannot be
found). In order that we do not need to use an ExplanatoryAssertionGroup
to list up all assumptions first and then possible again (in an
environment where not only failing assertions are shown), we need an
additional construct which allows us to report a summary => that's where
ISummaryAssertionGroupType shall help out.

- add ISummaryAssertionGroupType and two IBulletPointIdentifier, one for
  successful assertions and one for failing assertions
- introduce TextSummaryAssertionGroupFormatter
  - split up SingleAssertionGroupTypeFormatter, create a subclass which
    does the formatting of the children and a base class which does the
    rest