fix collecting feature assertions; access subject
Since feature assertions access the subject (via reflection) we have to
deal with this in contains.entries functions when collecting the
assertions. Ideally it takes an entry of the iterable as subject. Yet,
if the iterable is empty we have to catch the exception and show
something meaningful to the user

- let `subject` of ICollectingAssertionPlant be provided (by a lambda)
  - adopt IAtriumFactory accordingly
  - introduce ICollectingAssertionPlant.PlantHasNoSubjectException which
    should be thrown in case the provider cannot provide a `subject`
- `entries` function should deal with the situation that the `subject`
  might be absent when collecting assertions
  - extend specs accordingly