DownCastBuilder should not throw
If it throws we loose all information we collected so far. Thus it
should add an Assertion (which fails) to the original plant instead.
Furthermore it should return an plant which wraps assertions into
an ExplanatoryAssertionGroup. We also need to change API (cannot serve
a fluent API after such an assertion) but that will follow with the next
commit.

- create a IBaseAssertionPlant which already contains the addAssertion
  functions etc. and can be reused by IAssertionPlant and
  IAssertionPlantNullable
  - same for IBaseReportingAssertionPlant
  - modify BaseAssertionPlant and BaseReportingAssertionPlant
    accordingly
- introduce ExplanatoryAssertionPlant
- change DownCastBuilder
  - should no longer throw but add an assertion instead
  - returns the new ExplanatoryAssertionPlant instead
  - uses an DelegateAssertionChecker which passes on the assertions
  - change IAtriumFactory.newDownCastBuilder has to pass a subjectPlant
  - modify the specs accordingly
- remove nullRepresentation from IDownCastBuilder (will no longer report
  a failure)
  - add it to CommonFields instead
  - modify the specs accordingly