remove fluent API for narrowing functions
The problem with those functions (or rather the usage of the fluent api
in this case) are, that subsequent assertions depend on that the first
assertion (narrowing) holds. If not, we currently return an
ExplanatoryAssertionPlant which can only add more assertions as long as
they do not subject, only in the holds() fun respectively.
It is better if we remove those functions and only provide the overloads
with lambdas where the user can define subsequent assertions likewise.

- remove the overload for isA, isNotNull, Throwable.message (which
  uses implicit an isNotNull
  - keep the overload for toThrow but also return unit. I thank it is
    ok, if one only checks what kind of exception occurred without
    making subsequent assertions about this exception.
- modify DownCastBuilder accordingly
  - remove withLazyAssertions
  - expect the lambda as construction parameter
    - modify IAtriumFactory accordingly
- update the assertion functions (API) and specs accordingly