remove support for Locale no_...
As it seems there is a bug in ResourceBundle concerning resolving
no_NO_NY (caching destroys a deterministic Locale order). Due to this
I decided that I will not support no_... at all.
One can use either nb_... or nn_... and use one or the other as fallback
if desired. Thus:
- make sure a Translator throws an IllegalArgumentException in case the
  primary locale or one of its fallback starts with no_
- remove specialCaseNorwegian from LocaleOrderDecider
- remove TranslationSupplierNorwegianSpec incl. resources
-