introduce toBeWithErrorTolerance for floating point numbers
We will provide a sophisticated assertion builder later on (where one
could implement other error ranges, e.g. only positive error, relative
error and so on). Thus, we do not yet provide a shortcut function for
the infix API because it will probably not look different than the
sophisticated assertion builder. For now:

- introduce toBeWithErrorTolerance for cc-en_UK and cc-de_CH
- write spec
- include an exact check in the error message but only in the error
  message