deprecate toBe for BigDecimal, add isEqualIncludingScale
I think it is better if we force the user to use `isEqualIncludingScale`
instead of `toBe` if he really wants it.
The mistake of using `toBe` instead of `isNumericallyEqualTo` might not
pop up immediately and then it is kind of annoying that no one told the
developer "you should not use `toBe` for `BigDecimal` in most cases"
when the error/bug pops up. Thus, Atrium should say it ;)
Point to both `isNumericallyEqualTo` and to `isEqualIncludingScale` and
still give a hint if one decides to use `isEqualIncludingScale`. Why?
It could still be a mistake: e.g. one replaces deprecation too quickly;
does not really think about it or it might seem to be correct and in
in the future when the assertion does not hold anymore, then the
developer has forgotten about BigDecimal.scale and will search some time
why it does not hold as expected. I think it will be nice reminder.