move AsssertionGroup.Builder to own class
We already did the contrary once, move and rename AssertionGroupBuilder
to AssertionGroup.Builder.
I came to the conclusion that the AssertionGroupBuilder does way more
than the core api should provide. So this is the first step in a series
of commits. In the end we want to move the AssertionGroupBuilder to
atrium-assertions