introduce domain modules
Domain modules shall contain the domain logic of Atrium or in other
words, the implementation of the assertion functions as well as the
implementation of the sophisticated assertion builders which are both
used by the API modules.
As with core modules we want to provide the ability to exchange the
implementation to the users of Atrium. Thus:
- introduce atrium-domain-api
- atrium-domain-api-late-binding which is used for compileOnly
- atrium-domain-robstoll which merely has a dependency on:
  - atrium-domain-robstoll-lib which in turn contains the actual
    implementation