show exception message if it is a different type
Currently we do not show the `message` of a Throwable in Reporting
because we assume that most assertions about a thrown Throwable include
a feature assertion about its `message` and this already includes the
`message` of the Throwable. However, if another Exception is thrown than
the expected one, then we only see that the type does not much. The
message would be helpful in such cases. Thus:
- introduce method newExplanatoryWithHint for
  IAnyTypeTransformationFailureHandlers
- rename ExplanatoryTypeTransformationFailureHandler to
  ExplanatoryFailureHandler
  - move logic to a base class and reuse it
- introduce ExplanatoryFailureHandlerWithHint
  - introduce ThrowableThrownFailureHandler which makes use of it
    - use the new failure handler in ThrowableThrownAssertionCreator