re-add getters for CharSequence to be binary bc
Since the return types of the sophisticated builders have been more
specific in v0.6.0 than in the current implementation (we use interfaces
now), we have to re-add getters with the corresponding deprecated type
as return type. However, in order that we do not have to use the
deprecated types also as arguments we should relax the deprecated
classes. Thus:
- turn ContainsCheckerBuilder and CharSequenceContainsCheckerBuilder
  into interfaces
  - let the builders in the API implement
    CharSequenceContainsCheckerBuilder
- use the interfaces from package creating in the deprecated builders
- move searchbehaviours in package assertions from atirum-assertions to
  domain-api
- move ContainsChecker to atrium-assertions (is not required in
  domain-api)
- add an object CharSequenceContainsSearchersKt, we need it since we
  renamed charSequenceContainsSearchers.kt to
  charSequenceContainsCreators.kt
  - add all functions from charSequenceContainsCreators to it but with
    CharSequenceContainsCheckerBuilder as first argument and they need
    to be static (as it have been before)