introduce bc-tests with compilation
So far we have binary bc tests which are great but do not cover
everything because Kotlin supports inline functions. Therefore it might
be that a consumer is binary backward compatible but cannot longer
compile against the new version. Thus:

- add a corresponding task which retrieves the test sources, compiles
  them against the new API but the old spec and runs them
- rename current bc tasks to bbc (binary backward compatibility)
  - rename fun createBcTask to createBcTasks and reuse common logic

I was mistaking that gradle does not handle classifier tests correctly.
Defining a dependency on classifier tests for instance, does not add
a dependency to artifact without classifier (that's what I thought).
Thus:
- remove ...-tests projects again
  - create tests and testsources jar instead
- use tests and testsources jar in bc-test
- nameContainsNotValuesFun has to be public because we run the tests
  from a different module