add overload for property/returnValueOf for class reference
We want to provide assertion function writers the possibility to use
the overload which expects a class reference instead of a
bounded reference from the API. Most users will be familiar with the API
but not so much with the domain level. Hence, it is better if we
provide it in the API as well. Moreover, we should encourage users to
use it in context of a CollectingAssertionPlant