allow custom subject for AssertionCollector
This way it is easier to define custom feature assertions which involve
a transformation. For instance, turn an Iterable into a List and make
the assertion based on the List. So far we have created
CollectingAssertionPlants on our own and have done more or less what
the AssertionCollector is responsible for. Hence:
- expect a subjectProvider instead of an AssertionPlant
- still add an overload which expects plant and calls the other overload