allow translatable for property feature assertions
We already use a translatable but hide if from the API and turn String
into an Untranslatable. One should be able to use a Translatable for
property feature assertions. We can extend it to methods as well if we
have the need, so far I see only the need for properties.