move nullable overloads next to non-nullable and update KDoc
As we made a mistake in the past -- defined only nullable overloads or
only non-nullable overloads -- we should move the overloads next to each
other. This way it should be less likely that we make this mistake
again. We only do this for cc-de_CH, cc-en_GB and cc-infix-en_GB, or in
other words we do no longer care about cc-en_UK and cc-infix-en_UK.