add suffix to containsStrictlyNullable such as ...Values and ...Entries
containsStrictlyNullable is somewhat not complete; there is something
missing (the noun). I think it is better if we add a suffix to make it
clear otherwise one might come up with an interpretation like
`... contains (1 or null) and (2 or null)` for
`containsStrictlyNullable(1, 2)` => containsStrictlyNullableValues is
clearer.

For fluent APIS:
- rename containsStriclyNullable to containsStrictlyNullableValues/
  containsStrictlyNullableEntries
- introduce singular form containsStrictlyNullableValue/
  containsStrictlyNullableEntry

For infix API:
- use `containsStricly NullableValue` instead of
  containsStrictlyNullable