introduce ext String.format as replacement of String.Companion.format
I think that is better for two reasons. 1) there is a bug in Kotlin,
Intellij respectively which makes that extension functions of interface
Companion are not suggested in code completion -> makes working with it
cumbersome. 2) this way it is clearly distinguished from Java's
String.format and we can restrict the actual interface to some basic
functionality (JVM will provider richer functionality but this is then
considered an implementation Detail).

Moreover:
- introduce an invariant for TranslatableWithArgs, check that arguments
  are defined
- expect a List instead of an Array in TranslatableWithArgs
  - add overload for two arguments