fix getDefaultLocale() in JVM, should use toAtriumLocale
Currently it just passes the value of java.util.Locale to Atrium's
Locale which leads to errors since Atrium's Locale uses `null` instead
of empty strings to represent an absent configuration.

Moreover: so war we have used Class.name in all occasions in
DetailedObjectFormatter. This due to the fact that Java's primitive
`int` and the boxed type `Integer` are both represented as kotlin.Int
I still think it is a good idea if one sees the difference but since we
expect Any? for `format` we receive the boxed type in all cases. Hence
we can also just show the kotlin type in such cases. Which leads to the
following:
- make fullName return String instead of String?
- use qualifiedName if KClass is not a primitive type and use java.name
  as backup if qualifiedName returns null (which is the case for
  anonymous classes)
  - add this use case to ObjectFormatterSpec