reduce code duplication, use JvmMultifileClass
So far we have declared `expect` fun in the common module since we did
not want to include deprecated functionality in the JS platform. Thus
we have duplicated the implementation for JS and JVM. We can reduce
the code with the help of JvmMultifileClass.

This commit will not compile due to a bug in Kotlin, (file names have
to be different in common and jvm) but we make the commit nonetheless so
that git interprets it as a modification and not as new files. Fix
follows with the next commit.