fix nodejs gradle plugin nodeModulesDir has to point to parent dir
the name of the property is a bit misleading. Maybe the error does not
occur for all environment but in my case every second fail builds
because npm creates a link for node_modules/mocha to node_modules/mocha
which does not work of course. I don't encounter the problem when I use
the parent dir.