introduce partially fixed assertion group and use it for Map/List.get
showing whether a key exists in a Map or if the index is within its
bounds for List is kind of redundant information. It suffices that the
assertion group fails. Yet, we cannot use fixedClaimGroup since if
the pre-transformation (in this case extracting a feature) holds, then
the group can still fail if sub assertions fail. Thus

- introduce a group with a partially fixed part which forms an AND
  operation with the sub assertions.
- use it for List/Map.get
  - remove Description keys which are not longer required