Improve test descriptions to include function name and scenario
In the test reports, all of the tests have the similar names "does not
throw", "throws AssertionError" and as a result misses context of which
test does what.

This commit adds the function being tested to the test description and
also adds more description to tell what scenario leads to the observed
behavior.