change representation of thrown exception in case none is thrown
Previously we had `no exception thrown`, in this branch we are currently
showing `null` which is misleading. Thus we want `no exception thrown`
back. For this we need a way to override the representation of the
feature. Following the current approach of providing default arguments
to define Options is actually not ideal, because we could no longer have
a `val` for a feature such as size. Moreover, we already run into
ambiguity issues for the verbs, ideally we only have two versions,
fail-fast and non-fail-fast but not one for an act-lambda in addition.
For all this reasons:
- remove representation and options from the assertion verbs
- introduce RootExpect (part of #280)
- add an withOptions method to RootExpect so that one can re-define
  the ExpectOptions.
- introduce FeatureExpect (part of #280) together with FeatureOptions
  and the possibility to override them via withOptions
- introduce RootExpectConfig and FeatureExpectConfig on core-level
- introduce FeatureOptions analogous to ExpectOptions
- turn FeatureExtractorBuilder.OptionalRepresentationStep into
  an OptionsStep where one can pass/define FeatureOptions
- fix extractFeature: failure case should also show the intended
  feature in a group with a FeatureAssertionGroupType
- remove representation and ExpectOptions from verbs and simplify
  VerbSpec again