move parameterObjects to package creating, Windows-case-file-name-issue
The problem with keeping it in api.infix.en_GB is that we will generate
to class files with the same name in the sense of Windows
case insensitive file names. For instance
- open.class
- Open.class

That's a bummer. However, we want to use functions instead of the data
classes anyway (see #342) thus we are good to go