lower case all pseudo-keywords, move extension-keywords to normal API
By moving the keywords to atrium-api-infix-en_GB-common we hope that we
don't get into the situation where we would need the same keyword
also there (or in another extension) and thus run into name clashes.

For the same reason we move all parameter objects which are not
extension specific (e.g. PathWithCreator has to stay in jdk8 extension)
also to  atrium-api-infix-en_GB-common.

Moreover:
- make asPath infix with the help of `o`
- rename genericHelperFunctions.kt to helperFunctions.kt
  - move parameter objects which are not very specific to one type to
    the package creating (instead of e.g. creating.list)
- adjust module-info.java for API and extensions, it's totally wrong
  and Kotlin does not recognise it :(