avoid duplicate error message in IntelliJ console
by taking the advantage of the fact that both, the JUnit Jupiter runner
as well as Spek show Throwable.message next to do a printStacktrace
which invokes Throwable.getLocalizedMessage() which in turn calls
message usually. We tweak message to return an empty string and redirect
getLocalizedMessage() to super.message to work around the tweak.