move check-generateLogic-committed into own workflow due to github bug
Whatever I tried to do so far, as soon as I run one gradle script in
bash, the one in powershell does not behave as it would normally.
Thus moving the check into an own workflow.
The benefit, we only run it if something has changed in the input files
of the generation.