use Translator in ComponentFactoryContainer introduce ComponentFactory
ComponentFactory because we need a way to say we want kind of a
singleton which is re-used in this ComponentFactoryContainer.
We cannot / don't want to make guarantees about creating a singleton
only once, i.e. use computeIfAbsent or similar, because it would mean
we cannot have a singleton component which depends on another
singleton component.

We try at least to create a singleton as less as possible. For this to
work:
- introduce ComponentFactory (as already mentioned)
  - let ComponentFactoryContainer.getFactoryOrNull and
    getFactoryForChainedOrNull return a ComponentFactory instead of a
    simple function
- introduce ConcurrentMap as polyfill

moreover:
- deprecate get/setAtriumProperty and ReporterFactory
- deprecate all Translator and AssertionFormatter specific stuff in
  core-robstoll-lib and move to core
  - configure replacements in DefaultComponentFactoryContainer
  - get rid off most ReporterFactories only keep DefaultReporterFactory
    which we will remove with 0.17.0
- rename RemoveRunnerAtriumError to RemoveRunnerFromAtriumError
- move text specific AssertionFormatter to package `text` as preparation
  for further AssertionFormatters (like html)
- introduce RealTranslatorIntSpec which makes use of
  ComponentFactoryContainer instead of ReporterFactory