remove TODO on getImpl to replace it with components.
getImpl will be used in the future to create ProofFactory related types.
We will most likely rename it but it has a different purpose than
components. I think it will be much less likely that components are
actually exchanged where a user needs to define an own assertion verb to
do so. On the other hand, I could imagine that one wants to provide
an own implementation for a ProofFactory, e.g. to override one
particular expectation function, where with getImpl one is decoupled
from the assertion verb and instead provides an extension function
on the logic level.