apply publishing plugins, esp. tutteli-dokka only if env PUB is true
since the update to tutteli-dokka 5.0.0 we sometimes run into an
OutOfMemoryException in github_actions. the main change is that we now
create one dokkaHtml task per kotlin platform and my guess is, that they
have memory leaks (as dkoka had in the past). Hence, we only apply
dokka and tutteli-dokka when ENV PUB is given. On my local machine I
don't have any problems publishing.