update to Kotlin 1.9.10 no longer use translations_de_CH
kotlin 1.9.0 provides a new task
compileTestDevelopmentExecutableKotlinJs and it seems like this one
messes excluded dependencies up. It is still looking for
translations-en_GB even though we have excluded it.
As we are going to drop it in Atrium 1.3 any way, I think it is OK
if we already switch to translations_en_GB in api-infix