fail if `it` is used without `describe` in target JS
because our workaround (simplified):
```
it("our_test") { // generated by kotlin
  //turn test factory into describe/it
}
```
doesn't work otherwise, it looks like mocha ignores `it` within `it` and
only creates a new test if `describe` is used.

also, nonNullableCases cannot use testTemplateFactory as this would
create a RootExpectVerb, instead we need to use
turnTestNodesIntoExpectGrouping.

Last but not least, nonNullableCases, needs to expect a TestExecutable-
lambda and not TestFactoryBuilder-lambda